Explicitly install a recent version of transient package
Magit was updated and it requires a more recent version of transient.
However, the newer version of transient was not being installed
because transient is now built-in in Emacs (thus it was found and not
installed by elpaca). Now I'm explicitly installing transient from
its GitHub repository before installing magit.