Set word-wrap to t.
With this, `toggle-truncate-lines` will truncate on word boundaries. It
will then become visually similar to when `visual-line-mode` is used
instead of `toggle-truncate-lines`. However, note that there are
differences, such as `kill-line` killing only the visual line when
`visual-line-mode`  is used, while it kills the whole line (as if it
was not wrapped) when `toggle-truncate-lines` is used.